+++ src/commands/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod login;
 pub(crate) mod new;
 pub(crate) mod open;
 pub(crate) mod participate;
+pub(crate) mod retrieve_submission_summaries;
 pub(crate) mod retrieve_testcases;
 pub(crate) mod submit;
 pub(crate) mod test;